Add a common interpreter for units to extend
The interpreter can be extended by adding "keywords" (a mapping of a
string to function pointer). Keywords can operate on a piece of generic
data `T`, that is stored on the interpreter as the field `data`, but is
otherwise unused by the interpreter itself.

Simplifies adding an interpreter to a `unit` and adding common features
to all interpreters.