Fix InvalidArgs being raised by Mpu
Error occurs when both triggers and curves are used in a program and
only one of the Mpu's output functions are defined.

This is because the Mpu's output message wasn't being reset properly to
`MidiMessage::None`. So when one event set the message to a midi
ctrl or note event, the other function would assume a programmer error
if its function was then absent.

This also improves handling of the Mpu's functions from being absent.